New NeumorphicUIKitSilver.
This is a new version of NeumorphicUIKitSilver to be released.
Note: I am an inexperienced GitHub user. This is an attempt to perhaps upload it here.

☆ This was originally developed as a component library.
Improvements
All controls can now be adjusted for width and height.
Added Japanese for custom property name and description. Also, new properties have been added.
　The Switch and Button controls now have a property that allows switching to text.
Margins of each control are filled in to make layout easier.
New controls "ComboBox" and "Navigation" are added.
　ComboBox".... Combined with FluentUI drop-down.
　Navigation". . Slimmed down the Mobile Navigation Menu and enabled text input.

I thought it turned out to be quite easy to use, so I am releasing it to the public☺.